src/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
https://download.pytorch.org/whl/cu101/torch-1.7.1%2Bcu101-cp36-cp36m-linux_x86_64.whl
https://download.pytorch.org/whl/cu101/torchvision-0.8.2%2Bcu101-cp36-cp36m-linux_x86_64.whl
timm
# mmcv-full==1.3.0
# https://download.openmmlab.com/mmcv/dist/cu101/torch1.7.1/mmcv_full-1.2.7-cp36-cp36m-manylinux1_x86_64.whl
https://download.openmmlab.com/mmcv/dist/cu101/torch1.7.0/mmcv_full-1.3.0-cp36-cp36m-manylinux1_x86_64.whl
attrs
opencv-python-headless
matplotlib
numpy
terminaltables
ipython
pillow
